+++ client/src/components/EventCard/EventCard.css
@@ -1,11 +1,4 @@
-.card-image {
-  width: 100%;
-  height: 140px;
-  object-fit: cover;
-  clip-path: polygon(0 0, 100% 0, 100% 80%, 0% 100%);
-}
-
-.event-content {
+.card-content {
   display: flex;
   flex-direction: column;
   justify-content: flex-start;
@@ -35,7 +28,7 @@
   margin: 0;
 }
 
-.event-footer {
+.card-bottom {
   display: flex;
   justify-content: space-around;
   align-items: center;
@@ -55,4 +48,28 @@
   width: 16px;
   height: 16px;
   object-fit: contain;
+}
+
+@media (max-width: 768px) {
+  .event-content {
+    padding: 12px;
+  }
+
+  .event-title {
+    font-size: 1.1rem;
+  }
+
+  .event-style {
+    font-size: 0.95rem;
+  }
+
+  .event-footer {
+    flex-direction: column;
+    gap: 8px;
+    padding-bottom: 12px;
+  }
+
+  .event-time {
+    font-size: 1rem;
+  }
 }
