+++ client/src/types/EventType.ts
@@ -1,0 +1,11 @@
+export interface EventType {
+  id: number;
+  title: string;
+  start_at: string;
+  bar_id: number;
+  bar_name: string;
+  image: string;
+  music_group_id: number;
+  group_name: string;
+  music_style: string;
+}
